+++ src/main/java/svenhjol/charmony/travel_journal/client/features/travel_journal/Registers.java
@@ -1,6 +1,5 @@
 package svenhjol.charmony.travel_journal.client.features.travel_journal;
 
-import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import net.minecraft.client.KeyMapping;
@@ -8,14 +7,17 @@ import org.lwjgl.glfw.GLFW;
 import svenhjol.charmony.core.base.Setup;
 import svenhjol.charmony.core.client.ClientRegistry;
 import svenhjol.charmony.core.events.ClientLoginPlayerCallback;
+import svenhjol.charmony.core.events.PlayerTickCallback;
 import svenhjol.charmony.travel_journal.common.features.travel_journal.Networking.S2CSendBookmarkToPlayer;
 
 public final class Registers extends Setup<TravelJournal> {
     public KeyMapping openJournalKey;
     public KeyMapping makeBookmarkKey;
+    public final HudRenderer hudRenderer;
 
     public Registers(TravelJournal feature) {
         super(feature);
+        hudRenderer = new HudRenderer();
     }
 
     @Override
@@ -36,8 +38,8 @@ public final class Registers extends Setup<TravelJournal> {
                 "key.categories.misc"));
 
             ClientLoginPlayerCallback.EVENT.register(feature().handlers::clientLogin);
-            ClientTickEvents.END_CLIENT_TICK.register(feature().handlers::clientTick);
             HudRenderCallback.EVENT.register(feature().handlers::hudRender);
+            PlayerTickCallback.EVENT.register(feature().handlers::playerTick);
         };
     }
 }